add flexible item appenders.  current behavior seems fixed to top-down duplication of models in collection.
There is an index parameter that is an optional paramter to appendItem but it is *not* used when appendItem is called from the collection add (so it is only useful if you call appendItem directly which seems to go against the backbone way of model binding).

Outside of that, the renderCollection was calling them in order so you previously didn't have the ability to just override appendItem to change behavior as that would only work for dynamically added items - not items that were already in the collection when the view was rendered.

I do understand that you could add a comparator to the collection to change the order they are retrieved - but what if there is a single cache with different views?  Should the collection order be tied to the way that the DOM elements appear?