Honor the options.root scope
If we use view.$, then backbone will run queries scoped on the view. So
basically if you pass in an optional scope, say with options.root, then
it will just ignore it. However if you revert to $ than the options.root
scope will once again be honored. Take a look at the backbone source
[here](https://github.com/jashkenas/backbone/blob/1.0.0/backbone.js#L1006).

[Pull Request 229](https://github.com/walmartlabs/thorax/pull/229)