`detach` instead of `remove` old view in `LayoutView setView`
`$el.remove` also unbinds all events bound to the element, which is bad
when we're retaining the content view for reuse. This does not affect
views that expect the current behaviour of removing the element when
destroying, as that is still done by `_removeChild` if the retain count
of the view being removed is 0.

https://github.com/walmartlabs/thorax/issues/332