Collection item filter fix: parent var was shadowed
The fix for #306 introduced a new bug: the collection item filter
stopped working. I digged around, and notice that the error is because
when you try to verify the child is a direct child of the parent, it
uses the var `el`, which is also the name of the parent's element - so
the condition will never be true.
I changed the child's element var name to itemElement, which seems to
fix it.